Normalize line endings for non-Windows machines
I know this is mostly a Windows project, so feel free to deny this diff.
It should be a no-op for anyone running on Windows, and will only affect
folks on Linux or Mac.

Line endings are different on Windows and Macs, and this setting tells
git to handle the line endings, rather than just downloading what exists
in the repo. When I'm making a diff to the upstream project, this means
that my diffs will look like:
```
 if __name__ == "__main__":
+    print("hello")
     main()
```

instead of
```
 if __name__ == "__main__":^M
+    print("hello")^M
     main()^M

```

It's just a small quality-of-life change for non-Windows devs, and
shouldn't change anything for Windows devs.

No text was changed in this diff, other than removing some trailing
whitespace in seestar_run.py. Other than that, the only change is to the
invisible line ending characters (unfortunately, that makes it look like
every line was modified).