Renormalize line endings
This is a quality-of-life improvement for non-Windows users. Windows
defaults to CRLF line endings, Linux/Mac default to newlines (LF, aka
"\n"). This tells git to handle line endings based on the local system,
so diffs will look cleaner for Mac/Linux folks, and should be a no-op
for Windows folks.

This is where I found the commands to set this:
https://www.aleksandrhovhannisyan.com/blog/crlf-vs-lf-normalizing-line-endings-in-git/#a-simple-gitattributes-config